ebpf: split helper types and map code
