+++ pkg/ebpf/c/common/task.h
@@ -1,0 +1,178 @@
+#ifndef __TRACEE_TASK_H__
+#define __TRACEE_TASK_H__
+
+#ifndef CORE
+    #include <linux/sched.h>
+#else
+    // CO:RE is enabled
+    #include <vmlinux.h>
+#endif
+
+#include <bpf/bpf_core_read.h>
+#include <bpf/bpf_helpers.h>
+#include "common/arch.h"
+#include "common/common.h"
+#include "namespaces.h"
+
+static __always_inline int get_task_flags(struct task_struct *task)
+{
+    return READ_KERN(task->flags);
+}
+
+static __always_inline int get_task_syscall_id(struct task_struct *task)
+{
+    // There is no originated syscall in kernel thread context
+    if (get_task_flags(task) & PF_KTHREAD) {
+        return NO_SYSCALL;
+    }
+    struct pt_regs *regs = get_task_pt_regs(task);
+    return get_syscall_id_from_regs(regs);
+}
+
+static __always_inline u32 get_task_mnt_ns_id(struct task_struct *task)
+{
+    return get_mnt_ns_id(READ_KERN(task->nsproxy));
+}
+
+static __always_inline u32 get_task_pid_ns_for_children_id(struct task_struct *task)
+{
+    return get_pid_ns_for_children_id(READ_KERN(task->nsproxy));
+}
+
+static __always_inline u32 get_task_pid_ns_id(struct task_struct *task)
+{
+    unsigned int level = 0;
+    struct pid *pid = NULL;
+    struct pid_namespace *ns = NULL;
+
+#ifndef CORE
+    #if (LINUX_VERSION_CODE < KERNEL_VERSION(4, 19, 0) && !defined(RHEL_RELEASE_GT_8_0))
+    pid = READ_KERN(task->pids[PIDTYPE_PID].pid);
+    #else
+    pid = READ_KERN(task->thread_pid);
+    #endif
+#else
+    if (bpf_core_type_exists(struct pid_link)) {
+        struct task_struct___older_v50 *t = (void *) task;
+        pid = READ_KERN(t->pids[PIDTYPE_PID].pid);
+    } else {
+        pid = READ_KERN(task->thread_pid);
+    }
+#endif
+
+    level = READ_KERN(pid->level);
+    ns = READ_KERN(pid->numbers[level].ns);
+    return READ_KERN(ns->ns.inum);
+}
+
+static __always_inline u32 get_task_uts_ns_id(struct task_struct *task)
+{
+    return get_uts_ns_id(READ_KERN(task->nsproxy));
+}
+
+static __always_inline u32 get_task_ipc_ns_id(struct task_struct *task)
+{
+    return get_ipc_ns_id(READ_KERN(task->nsproxy));
+}
+
+static __always_inline u32 get_task_net_ns_id(struct task_struct *task)
+{
+    return get_net_ns_id(READ_KERN(task->nsproxy));
+}
+
+static __always_inline u32 get_task_cgroup_ns_id(struct task_struct *task)
+{
+    return get_cgroup_ns_id(READ_KERN(task->nsproxy));
+}
+
+static __always_inline u32 get_task_pid_vnr(struct task_struct *task)
+{
+    unsigned int level = 0;
+    struct pid *pid = NULL;
+
+#ifndef CORE
+    #if (LINUX_VERSION_CODE < KERNEL_VERSION(4, 19, 0) && !defined(RHEL_RELEASE_GT_8_0))
+    pid = READ_KERN(task->pids[PIDTYPE_PID].pid);
+    #else
+    pid = READ_KERN(task->thread_pid);
+    #endif
+#else
+    if (bpf_core_type_exists(struct pid_link)) {
+        struct task_struct___older_v50 *t = (void *) task;
+        pid = READ_KERN(t->pids[PIDTYPE_PID].pid);
+    } else {
+        pid = READ_KERN(task->thread_pid);
+    }
+#endif
+
+    level = READ_KERN(pid->level);
+    return READ_KERN(pid->numbers[level].nr);
+}
+
+static __always_inline u32 get_task_ns_pid(struct task_struct *task)
+{
+    return get_task_pid_vnr(task);
+}
+
+static __always_inline u32 get_task_ns_tgid(struct task_struct *task)
+{
+    struct task_struct *group_leader = READ_KERN(task->group_leader);
+    return get_task_pid_vnr(group_leader);
+}
+
+static __always_inline u32 get_task_ns_ppid(struct task_struct *task)
+{
+    struct task_struct *real_parent = READ_KERN(task->real_parent);
+    return get_task_pid_vnr(real_parent);
+}
+
+static __always_inline char *get_task_uts_name(struct task_struct *task)
+{
+    struct nsproxy *np = READ_KERN(task->nsproxy);
+    struct uts_namespace *uts_ns = READ_KERN(np->uts_ns);
+    return READ_KERN(uts_ns->name.nodename);
+}
+
+static __always_inline u32 get_task_ppid(struct task_struct *task)
+{
+    struct task_struct *parent = READ_KERN(task->real_parent);
+    return READ_KERN(parent->tgid);
+}
+
+static __always_inline u64 get_task_start_time(struct task_struct *task)
+{
+    return READ_KERN(task->start_time);
+}
+
+static __always_inline u32 get_task_host_pid(struct task_struct *task)
+{
+    return READ_KERN(task->pid);
+}
+
+static __always_inline u32 get_task_host_tgid(struct task_struct *task)
+{
+    return READ_KERN(task->tgid);
+}
+
+static __always_inline struct task_struct *get_parent_task(struct task_struct *task)
+{
+    return READ_KERN(task->real_parent);
+}
+
+static __always_inline u32 get_task_exit_code(struct task_struct *task)
+{
+    return READ_KERN(task->exit_code);
+}
+
+static __always_inline int get_task_parent_flags(struct task_struct *task)
+{
+    struct task_struct *parent = READ_KERN(task->real_parent);
+    return get_task_flags(parent);
+}
+
+static __always_inline const struct cred *get_task_real_cred(struct task_struct *task)
+{
+    return READ_KERN(task->real_cred);
+}
+
+#endif
